# apps/ios/LorealClienteling/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/products": {
            "get": {
                "operationId": "ProductsController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Products"
                ]
            },
            "post": {
                "operationId": "ProductsController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Products"
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "operationId": "ProductsController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Products"
                ]
            },
            "patch": {
                "operationId": "ProductsController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Products"
                ]
            }
        },
        "/products/{id}/availability": {
            "get": {
                "operationId": "ProductsController_getAvailability",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Products"
                ]
            }
        },
        "/products/{id}/availability/{storeId}": {
            "patch": {
                "operationId": "ProductsController_updateAvailability",
                "parameters": [
                    {
                        "name": "storeId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAvailabilityDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Products"
                ]
            }
        },
...
        "/customers/{customerId}/consents": {
            "get": {
                "operationId": "ConsentsController_findByCustomer",
                "parameters": [
                    {
                        "name": "customerId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Consents"
                ]
            },
            "post": {
                "operationId": "ConsentsController_grant",
                "parameters": [
                    {
                        "name": "customerId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GrantConsentDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Consents"
                ]
            }
        },
        "/customers/{customerId}/consents/{type}": {
            "delete": {
                "operationId": "ConsentsController_revoke",
                "parameters": [
                    {
                        "name": "type",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customerId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Consents"
                ]
            }
        },
...
        "/communications": {
            "get": {
                "operationId": "CommunicationsController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Communications"
                ]
            },
            "post": {
                "operationId": "CommunicationsController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCommunicationDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Communications"
                ]
            }
        },
        "/customers/{customerId}/communications": {
            "get": {
                "operationId": "CommunicationsController_findByCustomer",
                "parameters": [
                    {
                        "name": "customerId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Communications"
                ]
            }
        },
        "/communications/templates": {
            "get": {
                "operationId": "CommunicationsController_findTemplates",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Communications"
                ]
            },
            "post": {
                "operationId": "CommunicationsController_createTemplate",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTemplateDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Communications"
                ]
            }
        },
        "/communications/templates/{id}": {
            "patch": {
                "operationId": "CommunicationsController_updateTemplate",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTemplateDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Communications"
                ]
            }
        },
        "/communications/{id}/tracking": {
            "patch": {
                "operationId": "CommunicationsController_updateTracking",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTrackingDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Communications"
                ]
            }
        },
...
    },
    "info": {
        "title": "L'Or\u00e9al Clienteling API",
        "description": "API for L'Or\u00e9al beauty advisor clienteling platform",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "http://localhost:3001",
            "description": "Development"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateProductDto": {
                "type": "object",
                "properties": {
                    "sku": {
                        "type": "string",
                        "example": "SKU-001",
                        "minLength": 1,
                        "maxLength": 50
                    },
                    "name": {
                        "type": "string",
                        "example": "Revitalift S\u00e9rum",
                        "minLength": 1,
                        "maxLength": 200
                    },
                    "brandId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "skincare",
                            "makeup",
                            "fragrance"
                        ],
                        "example": "skincare"
                    },
                    "subcategory": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 2000
                    },
                    "price": {
                        "type": "number",
                        "example": 1299,
                        "minimum": 0
                    },
                    "estimatedDurationDays": {
                        "type": "number",
                        "example": 90
                    }
                },
                "required": [
                    "sku",
                    "name",
                    "brandId",
                    "category",
                    "price"
                ]
            },
            "UpdateProductDto": {
                "type": "object",
                "properties": {
                    "sku": {
                        "type": "string",
                        "example": "SKU-001",
                        "minLength": 1,
                        "maxLength": 50
                    },
                    "name": {
                        "type": "string",
                        "example": "Revitalift S\u00e9rum",
                        "minLength": 1,
                        "maxLength": 200
                    },
                    "brandId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "skincare",
                            "makeup",
                            "fragrance"
                        ],
                        "example": "skincare"
                    },
                    "subcategory": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 2000
                    },
                    "price": {
                        "type": "number",
                        "example": 1299,
                        "minimum": 0
                    },
                    "estimatedDurationDays": {
                        "type": "number",
                        "example": 90
                    }
                }
            },
            "UpdateAvailabilityDto": {
                "type": "object",
                "properties": {
                    "stockStatus": {
                        "type": "string",
                        "enum": [
                            "available",
                            "low",
                            "out_of_stock"
                        ],
                        "example": "available"
                    }
                },
                "required": [
                    "stockStatus"
                ]
            },
...
            "GrantConsentDto": {
                "type": "object",
                "properties": {
                    "customerId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "privacy_notice",
                            "marketing_sms",
                            "marketing_email",
                            "marketing_whatsapp"
                        ],
                        "example": "privacy_notice"
                    },
                    "version": {
                        "type": "string",
                        "example": "1.0",
                        "minLength": 1,
                        "maxLength": 20
                    },
                    "source": {
                        "type": "string",
                        "example": "app",
                        "maxLength": 100
                    }
                },
                "required": [
                    "customerId",
                    "type",
                    "version",
                    "source"
                ]
            },
...
            "CreateCommunicationDto": {
                "type": "object",
                "properties": {
                    "customerId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "channel": {
                        "type": "string",
                        "enum": [
                            "whatsapp",
                            "sms",
                            "email"
                        ],
                        "example": "whatsapp"
                    },
                    "templateId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "subject": {
                        "type": "string",
                        "maxLength": 200
                    },
                    "body": {
                        "type": "string",
                        "example": "Hola Mar\u00eda, tu cita est\u00e1 confirmada.",
                        "minLength": 1,
                        "maxLength": 5000
                    },
                    "followupType": {
                        "type": "string",
                        "enum": [
                            "3_months",
                            "6_months",
                            "birthday",
                            "replenishment",
                            "special_event",
                            "custom"
                        ],
                        "example": "3_months"
                    }
                },
                "required": [
                    "customerId",
                    "channel",
                    "body",
                    "followupType"
                ]
            },
            "CreateTemplateDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Seguimiento 3 meses",
                        "minLength": 1,
                        "maxLength": 200
                    },
                    "brandId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "channel": {
                        "type": "string",
                        "enum": [
                            "whatsapp",
                            "sms",
                            "email"
                        ],
                        "example": "whatsapp"
                    },
                    "followupType": {
                        "type": "string",
                        "enum": [
                            "3_months",
                            "6_months",
                            "birthday",
                            "replenishment",
                            "special_event",
                            "custom"
                        ],
                        "example": "3_months"
                    },
                    "body": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 5000
                    }
                },
                "required": [
                    "name",
                    "channel",
                    "followupType",
                    "body"
                ]
            },
            "UpdateTemplateDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Seguimiento 3 meses",
                        "minLength": 1,
                        "maxLength": 200
                    },
                    "brandId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "channel": {
                        "type": "string",
                        "enum": [
                            "whatsapp",
                            "sms",
                            "email"
                        ],
                        "example": "whatsapp"
                    },
                    "followupType": {
                        "type": "string",
                        "enum": [
                            "3_months",
                            "6_months",
                            "birthday",
                            "replenishment",
                            "special_event",
                            "custom"
                        ],
                        "example": "3_months"
                    },
                    "body": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 5000
                    }
                }
            },
            "UpdateTrackingDto": {
                "type": "object",
                "properties": {
                    "deliveredAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "readAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "respondedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            }
        }
    }
}